Conflicts:
	c/Makefile

+++ c/ntdrivers/parport_false-unreach-call.i.cil.c
@@ -1934,7 +1934,7 @@ struct _SCSI_REQUEST_BLOCK;
 #pragma warning(push)
 #pragma warning(disable:4035)
 #pragma warning(pop)
-extern void *malloc(unsigned long sz ) ;
+extern void *malloc(unsigned int sz ) ;
 extern void *memcpy(void * , void const   * , size_t  ) ;
 extern void *memset(void * , int  , size_t  ) ;
 extern void *memmove(void * , void const   * , size_t  ) ;
